rename review actions
